Fixed system test and ProofOfWorkService

+++ core/src/main/java/ch/dissem/bitmessage/ProofOfWorkService.java
@@ -1,14 +1,15 @@
 package ch.dissem.bitmessage;
 
-import ch.dissem.bitmessage.entity.BitmessageAddress;
-import ch.dissem.bitmessage.entity.ObjectMessage;
-import ch.dissem.bitmessage.entity.Plaintext;
-import ch.dissem.bitmessage.entity.PlaintextHolder;
+import ch.dissem.bitmessage.entity.*;
+import ch.dissem.bitmessage.entity.payload.Broadcast;
+import ch.dissem.bitmessage.entity.payload.Msg;
+import ch.dissem.bitmessage.entity.payload.ObjectPayload;
 import ch.dissem.bitmessage.entity.payload.Pubkey;
 import ch.dissem.bitmessage.ports.Cryptography;
 import ch.dissem.bitmessage.ports.MessageRepository;
 import ch.dissem.bitmessage.ports.ProofOfWorkEngine;
 import ch.dissem.bitmessage.ports.ProofOfWorkRepository;
+import ch.dissem.bitmessage.ports.ProofOfWorkRepository.Item;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -16,7 +17,7 @@ import java.util.List;
 
 import static ch.dissem.bitmessage.InternalContext.NETWORK_EXTRA_BYTES;
 import static ch.dissem.bitmessage.InternalContext.NETWORK_NONCE_TRIALS_PER_BYTE;
-import static ch.dissem.bitmessage.utils.Singleton.security;
+import static ch.dissem.bitmessage.utils.Singleton.cryptography;
 
 /**
  * @author Christian Basler
@@ -35,7 +36,7 @@ public class ProofOfWorkService implements ProofOfWorkEngine.Callback, InternalC
 
         LOG.info("Doing POW for " + items.size() + " tasks.");
         for (byte[] initialHash : items) {
-            ProofOfWorkRepository.Item item = powRepo.getItem(initialHash);
+            Item item = powRepo.getItem(initialHash);
             cryptography.doProofOfWork(item.object, item.nonceTrialsPerByte, item.extraBytes, this);
         }
     }
@@ -59,25 +60,50 @@ public class ProofOfWorkService implements ProofOfWorkEngine.Callback, InternalC
         cryptography.doProofOfWork(object, nonceTrialsPerByte, extraBytes, this);
     }
 
+    public void doProofOfWorkWithAck(Plaintext plaintext, long expirationTime) {
+        final ObjectMessage ack = plaintext.getAckMessage();
+        Item item = new Item(ack, NETWORK_NONCE_TRIALS_PER_BYTE, NETWORK_EXTRA_BYTES,
+                expirationTime, plaintext);
+        powRepo.putObject(item);
+        cryptography.doProofOfWork(ack, NETWORK_NONCE_TRIALS_PER_BYTE, NETWORK_EXTRA_BYTES, this);
+    }
+
     @Override
     public void onNonceCalculated(byte[] initialHash, byte[] nonce) {
-        ObjectMessage object = powRepo.getItem(initialHash).object;
-        object.setNonce(nonce);
-        Plaintext plaintext = messageRepo.getMessage(initialHash);
-        if (plaintext != null) {
-            plaintext.setInventoryVector(object.getInventoryVector());
-            ctx.getLabeler().markAsSent(plaintext);
-            messageRepo.save(plaintext);
+        Item item = powRepo.getItem(initialHash);
+        if (item.message == null) {
+            ObjectMessage object = powRepo.getItem(initialHash).object;
+            object.setNonce(nonce);
+            Plaintext plaintext = messageRepo.getMessage(initialHash);
+            if (plaintext != null) {
+                plaintext.setInventoryVector(object.getInventoryVector());
+                ctx.getLabeler().markAsSent(plaintext);
+                messageRepo.save(plaintext);
+            }
+            ctx.getInventory().storeObject(object);
+            powRepo.removeObject(initialHash);
+            ctx.getNetworkHandler().offer(object.getInventoryVector());
+        } else {
+            item.message.getAckMessage().setNonce(nonce);
+            final ObjectMessage object = new ObjectMessage.Builder()
+                    .stream(item.message.getStream())
+                    .expiresTime(item.expirationTime)
+                    .payload(new Msg(item.message))
+                    .build();
+            if (object.isSigned()) {
+                object.sign(item.message.getFrom().getPrivateKey());
+            }
+            if (object.getPayload() instanceof Encrypted) {
+                object.encrypt(item.message.getTo().getPubkey());
+            }
+            doProofOfWork(item.message.getTo(), object);
         }
-        ctx.getInventory().storeObject(object);
-        powRepo.removeObject(initialHash);
-        ctx.getNetworkHandler().offer(object.getInventoryVector());
     }
 
     @Override
     public void setContext(InternalContext ctx) {
         this.ctx = ctx;
-        this.cryptography = security();
+        this.cryptography = cryptography();
         this.powRepo = ctx.getProofOfWorkRepository();
         this.messageRepo = ctx.getMessageRepository();
     }
